Rename component file names
